add function comment. add a client wait execution function for parser result for local test.
